+++ Changelog.md
@@ -1,3 +1,8 @@
+### 0.2.0.0
+
+- Render substitutions using Pandoc's writers.
+- Put explicitly in docs that only HTML and LaTeX formats are supported.
+
 ### 0.1.1.0
 
 - Added support for Pandoc 2.0.